content_writer/pyproject.toml
[project]
name = "content-writer"
version = "2.0.0"
description = "Hybrid OpenAI/Ollama content generator for audispot254"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
    "youtube-transcript-api>=0.6.0",
]
